@@ -12,16 +12,15 @@ import Darwin.C
 #endif
 
 /// A type that represents the structure and behavior of an app.
-public protocol App {
+@MainActor @preconcurrency
+public protocol App: Sendable {
 
     /// Creates an instance of the app using the body that you define for its content.
-    @MainActor @preconcurrency
     init()
 
     associatedtype Content: AppScene
     
     /// Main scene in your app.
-    @MainActor @preconcurrency
     var scene: Content { get }
 }
 
@@ -32,10 +31,9 @@ public extension App {
     }
     
     // Initializes and runs the app.
-    static func main() throws {
+    static func main() async throws {
         let appContext = try AppContext<Self>()
-        try appContext.setup()
-
+        try await appContext.setup()
         try appContext.runApplication()
     }
 }
