@@ -179,8 +179,8 @@ public enum StencilOperation: UInt, Codable {
 // MARK: - Others -
 
 /// The geometric primitive type for drawing commands.
-public enum IndexPrimitive: UInt8, Codable {
-    
+public enum IndexPrimitive: UInt8, Codable, Sendable {
+
     /// For every separate set of three vertices, rasterize a triangle. If the number of vertices is not a multiple of three, either one or two vertices is ignored.
     case triangle
     
@@ -198,8 +198,8 @@ public enum IndexPrimitive: UInt8, Codable {
 }
 
 /// Types of actions performed for an attachment at the start of a rendering pass.
-public enum AttachmentLoadAction: Codable {
-    
+public enum AttachmentLoadAction: Codable, Sendable {
+
     /// The GPU writes a value to every pixel in the attachment at the start of the render pass.
     case clear
     
@@ -211,8 +211,8 @@ public enum AttachmentLoadAction: Codable {
 }
 
 /// Types of actions performed for an attachment at the end of a rendering pass.
-public enum AttachmentStoreAction: Codable {
-    
+public enum AttachmentStoreAction: Codable, Sendable {
+
     /// The GPU stores the rendered contents to the texture.
     case store
     
@@ -221,8 +221,8 @@ public enum AttachmentStoreAction: Codable {
 }
 
 /// Options used to specify how a sample compare operation should be performed on a depth texture.
-public enum CompareOperation: UInt, Codable {
-    
+public enum CompareOperation: UInt, Codable, Sendable {
+
     /// A new value never passes the comparison test.
     case never
     
@@ -250,7 +250,7 @@ public enum CompareOperation: UInt, Codable {
 }
 
 /// The data formats that describe the organization and characteristics of individual pixels in a texture.
-public enum PixelFormat: Codable {
+public enum PixelFormat: Codable, Sendable {
     
     /// You cannot create a texture with this value.
     case none
