@@ -14,7 +14,7 @@ public enum TriangleFillMode {
 
 /// This protocol describe interface for GPU.
 protocol RenderBackend: AnyObject {
-    
+
     /// Returns current frame index. Min value 0, Max value is equal ``RenderEngine/Configuration/maxFramesInFlight`` value.
     var currentFrameIndex: Int { get }
 
@@ -27,21 +27,21 @@ protocol RenderBackend: AnyObject {
     /// Register a new render window for render backend.
     /// Window in this case is entity that managed a drawables (aka swapchain).
     /// - Throws: Throw error if something went wrong.
-    func createWindow(_ windowId: UIWindow.ID, for surface: RenderSurface, size: SizeInt) throws
+    @MainActor func createWindow(_ windowId: UIWindow.ID, for surface: RenderSurface, size: SizeInt) throws
 
     /// Resize registred render window.
     /// - Throws: Throw error if window is not registred.
-    func resizeWindow(_ windowId: UIWindow.ID, newSize: SizeInt) throws
-    
+    @MainActor func resizeWindow(_ windowId: UIWindow.ID, newSize: SizeInt) throws
+
     /// Destroy render window from render backend.
     /// - Throws: Throw error if window is not registred.
-    func destroyWindow(_ windowId: UIWindow.ID) throws
-    
+    @MainActor func destroyWindow(_ windowId: UIWindow.ID) throws
+
     /// Begin rendering a frame for all windows.
-    func beginFrame() throws
-    
+    @MainActor func beginFrame() throws
+
     /// Release any data associated with the current frame.
-    func endFrame() throws
+    @MainActor func endFrame() throws
 }
 
 /// The GPU device instance resposible for rendering and computing.
