@@ -10,7 +10,7 @@ public struct SpritePlugin: ScenePlugin {
     
     public init() {}
     
-    public func setup(in scene: Scene) async {
+    public func setup(in scene: Scene) {
         scene.addSystem(ExtractSpriteSystem.self)
     }
 }
