@@ -53,7 +53,7 @@ public struct SpriteRenderSystem: RenderSystem {
     // MARK: - Private
 
     // swiftlint:disable:next function_body_length
-    private func draw(
+    @MainActor private func draw(
         extractedSprites: [ExtractedSprite],
         visibleEntities: VisibleEntities,
         renderItems: inout RenderItems<Transparent2DRenderItem>
@@ -223,7 +223,7 @@ public struct ExtractSpriteSystem: System {
 
     public init(scene: Scene) { }
 
-    public func update(context: UpdateContext) async {
+    public func update(context: UpdateContext) {
         let extractedEntity = EmptyEntity()
         var extractedSprites = ExtractedSprites(sprites: [])
 
@@ -246,6 +246,8 @@ public struct ExtractSpriteSystem: System {
         }
 
         extractedEntity.components += extractedSprites
-        await Application.shared.renderWorld.addEntity(extractedEntity)
+        Task {
+            await Application.shared.renderWorld.addEntity(extractedEntity)
+        }
     }
 }
