@@ -16,11 +16,10 @@ import Collections
 /// Each ``Entity`` has a set of components. Each component can have up to one instance of each
 /// component type. Entity components can be created, updated, removed, and queried using a given World.
 /// - Warning: Still work in progress.
+@MainActor
 public final class World {
 
     private var records: OrderedDictionary<Entity.ID, EntityRecord> = [:]
-    
-    let lock = NSRecursiveLock()
 
     internal private(set) var removedEntities: Set<Entity.ID> = []
     internal private(set) var addedEntities: Set<Entity.ID> = []
@@ -41,9 +40,6 @@ public final class World {
     /// Get all entities in world.
     /// - Complexity: O(n)
     public func getEntities() -> [Entity] {
-        lock.lock()
-        defer { lock.unlock() }
-        
         return self.records.values.elements
             .map { record in
                 let archetype = self.archetypes[record.archetypeId]!
@@ -55,9 +51,6 @@ public final class World {
     /// Get entity by identifier.
     /// - Complexity: O(1)
     public func getEntityByID(_ entityID: Entity.ID) -> Entity? {
-        lock.lock()
-        defer { lock.unlock() }
-        
         guard let record = self.records[entityID] else {
             return nil
         }
@@ -69,9 +62,6 @@ public final class World {
     /// Get entity by name.
     /// - Complexity: O(n)
     func getEntityByName(_ name: String) -> Entity? {
-        lock.lock()
-        defer { lock.unlock() }
-        
         for arch in archetypes {
             if let ent = arch.entities.first(where: { $0.name == name }) {
                 return ent
@@ -84,9 +74,6 @@ public final class World {
     // FIXME: Can crash if we change components set during runtime
     /// Append entity to world. Entity will be added when `tick()` called.
     func appendEntity(_ entity: Entity) {
-        lock.lock()
-        defer { lock.unlock() }
-
         for (identifier, component) in entity.components.buffer {
             if let script = component as? ScriptableComponent {
                 self.addScript(script, entity: entity.id, identifier: identifier)
@@ -100,16 +87,10 @@ public final class World {
     }
     
     func removeEntity(_ entity: Entity) {
-        lock.lock()
-        defer { lock.unlock() }
-        
         self.removeEntityRecord(entity.id)
     }
     
     func removeEntityOnNextTick(_ entity: Entity, recursively: Bool = false) {
-        lock.lock()
-        defer { lock.unlock() }
-        
         guard self.records[entity.id] != nil else { 
             return
         }
@@ -130,9 +111,6 @@ public final class World {
     }
 
     private func removeEntityRecord(_ entity: Entity.ID) {
-        lock.lock()
-        defer { lock.unlock() }
-        
         guard let record = self.records[entity] else {
             return
         }
@@ -251,9 +229,6 @@ public final class World {
     // MARK: - Components Delegate
 
     func entity(_ entity: Entity, didAddComponent component: Component, with identifier: ComponentId) {
-        lock.lock()
-        defer { lock.unlock() }
-
         if let script = component as? ScriptableComponent {
             self.addScript(script, entity: entity.id, identifier: identifier)
         }
@@ -265,9 +240,6 @@ public final class World {
     }
 
     func entity(_ entity: Entity, didUpdateComponent component: Component, with identifier: ComponentId) {
-        lock.lock()
-        defer { lock.unlock() }
-
         if let script = component as? ScriptableComponent {
             self.addScript(script, entity: entity.id, identifier: identifier)
         }
@@ -280,9 +252,6 @@ public final class World {
     }
 
     func entity(_ entity: Entity, didRemoveComponent component: Component.Type, with identifier: ComponentId) {
-        lock.lock()
-        defer { lock.unlock() }
-
         EventManager.default.send(ComponentEvents.WillRemove(componentType: component, entity: entity))
 
         if component is ScriptableComponent.Type {
