@@ -10,7 +10,7 @@ public struct Physics2DPlugin: ScenePlugin {
     
     public init() {}
     
-    public func setup(in scene: Scene) async {
+    public func setup(in scene: Scene) {
         /// We have physics world as an entity, because it's more flexible solution then store
         /// physics world in the scene object.
         let physicsWorldEntity = Entity(name: "PhysicsWorld2D")
@@ -21,8 +21,10 @@ public struct Physics2DPlugin: ScenePlugin {
         scene.addEntity(physicsWorldEntity)
         scene.addSystem(DebugPhysicsExctract2DSystem.self)
         scene.addSystem(Physics2DSystem.self)
-        
-        await Application.shared.renderWorld.addSystem(Physics2DDebugDrawSystem.self)
+
+        Task {
+            await Application.shared.renderWorld.addSystem(Physics2DDebugDrawSystem.self)
+        }
     }
 }
 
