@@ -8,25 +8,24 @@
 /// Contains base configuration for any scene in the game.
 /// This plugins will applied for each scene in Ada application and should be passed once per scene.
 struct DefaultScenePlugin: ScenePlugin {
-    func setup(in scene: Scene) async {
+    func setup(in scene: Scene) {
         // Add base systems
         scene.addSystem(ScriptComponentUpdateSystem.self)
 
         // Setup render
-        await scene.addPlugin(VisibilityPlugin())
-        await scene.addPlugin(CameraPlugin())
+        scene.addPlugin(VisibilityPlugin())
+        scene.addPlugin(CameraPlugin())
 
-        await scene.addPlugin(SpritePlugin())
-        await scene.addPlugin(Mesh2DPlugin())
-        await scene.addPlugin(Text2DPlugin())
-        await scene.addPlugin(AudioPlugin())
-        await scene.addPlugin(UIPlugin())
+        scene.addPlugin(SpritePlugin())
+        scene.addPlugin(Mesh2DPlugin())
+        scene.addPlugin(Text2DPlugin())
+        scene.addPlugin(AudioPlugin())
+        scene.addPlugin(UIPlugin())
 
         // Setup Physics
-        await scene.addPlugin(Physics2DPlugin())
-
-        await scene.addPlugin(TransformPlugin())
-        await scene.addPlugin(TileMapPlugin())
+        scene.addPlugin(Physics2DPlugin())
+        scene.addPlugin(TransformPlugin())
+        scene.addPlugin(TileMapPlugin())
     }
 }
 
