@@ -20,8 +20,8 @@ public struct Text2DLayoutSystem: System {
     
     public init(scene: Scene) {}
     
-    public func update(context: UpdateContext) async {
-        await context.scene.performQuery(Self.textComponents).concurrent.forEach { entity in
+    public func update(context: UpdateContext) {
+        context.scene.performQuery(Self.textComponents).forEach { entity in
             let (text, visibility) = entity.components[Text2DComponent.self, Visibility.self]
             
             if !visibility.isVisible {
