@@ -10,7 +10,7 @@ public struct Text2DPlugin: ScenePlugin {
     
     public init() {}
     
-    public func setup(in scene: Scene) async {
+    public func setup(in scene: Scene) {
         scene.addSystem(ExctractTextSystem.self)
         scene.addSystem(Text2DLayoutSystem.self)
     }
