@@ -39,7 +39,7 @@ class DrawingGroupViewNode: ViewModifierNode {
     }
 
     override func createLayer() -> UILayer? {
-        let layer = UILayer(frame: self.frame) { [weak self] context, size in
+        let layer = UILayer(frame: self.frame) { [weak self] context, _ in
             guard let self else {
                 return
             }
